some changes to PhotoService

+++ src/main/java/ru/skypro/homework/service/PhotoService.java
@@ -112,6 +112,8 @@ public class PhotoService {
     }
 
     private void deletePreviousPhotoInDirectory(Photo photo) throws IOException {
-        Files.deleteIfExists(photo.getFilePath().toAbsolutePath().toFile().toPath());
+        if (photo != null) {
+            Files.deleteIfExists(photo.getFilePath().toAbsolutePath().toFile().toPath());
+        }
     }
 }
